Major project rework
-Converted all projects to .NET 5
-Migrated most functionality to DidasUtils
-Moved old messages in Primes.Common.Net to .old files
-Moved versions.xml to Resources and discarded unused .xml files
-PrimesUpdater renamed to PrimesInstaller in the process of NET version changing
-PrimesUpdater now checks if working on Windows as other OSes are not supported yet
-Moved incomplete BatchDownloader functionality of Primes to .old file
-Renamed conflicting files/classes in Primes.Common and DidasUtils
-Removed BatchDistributer

+++ PrimesInstaller/CmdHelper.cs
@@ -3,65 +3,93 @@ using System.IO;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
 
-namespace Primes.Updater
+namespace Primes.Installer
 {
     public static class CmdHelper
     {
         public static string BuildSelfUpdateCMD(string uncompressPath, string installPath, string execToStart)
         {
-            return $"ping -n 3 127.0.0.1>nul\n" +
+            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+            {
+                return $"ping -n 3 127.0.0.1>nul\n" +
                     $"taskkill -f -im 7za.exe\n" +
                     $"del /f /q /s \"{installPath}\\*.*\"\n" +
                     $"xcopy \"{uncompressPath}\\*.*\" \"{installPath}\"\n" +
                     $"start \"PrimesUpdater.exe\" \"{installPath}\\{execToStart}\" -ns\n" +
                     $"ping -n 5 127.0.0.1>nul\n" +
                     $"exit";
+            }
+            else
+            {
+                throw new NotImplementedException("Installation on Linux/OSX is not supported yet");
+            }
         }
         public static string BuildUpdateCMD(string uncompressPath, string installPath, string containingDirName)
         {
-            return $"@ping -n 3 127.0.0.1>nul\n" +
+            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+            {
+                return $"@ping -n 3 127.0.0.1>nul\n" +
                     $"@taskkill -f -im 7za.exe\n" +
                     $"for /d %%a in (\"{installPath}\\*\") do rd /S /Q \"%%a\"\n" +
                     $"for %%a in (\"{installPath}\") do if /i not \"%%~nxa\"==\"settings.set\" del /q \"%%a\"\n" +
                     $"xcopy \"{uncompressPath}\\{containingDirName}\\bin\\*.*\" \"{installPath}\"\n" +
                     $"@ping -n 3 127.0.0.1>nul\n" +
                     $"exit";
+            }
+            else
+            {
+                throw new NotImplementedException("Installation on Linux/OSX is not supported yet");
+            }
         }
         public static void LaunchSelfUpdateCMD(string cmdCode, string tmpDir)
         {
-            string programPath = Path.Combine(tmpDir, "update.bat");
-            File.WriteAllText(programPath, cmdCode);
-
-            ProcessStartInfo startInfo = new ProcessStartInfo()
+            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
             {
-                FileName = programPath
-            };
+                string programPath = Path.Combine(tmpDir, "update.bat");
+                File.WriteAllText(programPath, cmdCode);
 
-            Process updater = new Process
+                ProcessStartInfo startInfo = new()
+                {
+                    FileName = programPath
+                };
+
+                Process updater = new()
+                {
+                    StartInfo = startInfo
+                };
+                updater.Start();
+
+                Environment.Exit(0);
+            }
+            else
             {
-                StartInfo = startInfo
-            };
-            updater.Start();
-
-            Environment.Exit(0);
+                throw new NotImplementedException("Installation on Linux/OSX is not supported yet");
+            }
         }
         public static void RunUpdateCMD(string cmdCode, string tmpDir)
         {
-            string programPath = Path.Combine(tmpDir, "update.bat");
-            File.WriteAllText(programPath, cmdCode);
-
-            ProcessStartInfo startInfo = new ProcessStartInfo()
+            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
             {
-                FileName = programPath
-            };
+                string programPath = Path.Combine(tmpDir, "update.bat");
+                File.WriteAllText(programPath, cmdCode);
 
-            Process updater = new Process
+                ProcessStartInfo startInfo = new()
+                {
+                    FileName = programPath
+                };
+
+                Process updater = new()
+                {
+                    StartInfo = startInfo
+                };
+                updater.Start();
+
+                updater.WaitForExit();
+            }
+            else
             {
-                StartInfo = startInfo
-            };
-            updater.Start();
-
-            updater.WaitForExit();
+                throw new NotImplementedException("Installation on Linux/OSX is not supported yet");
+            }
         }
     }
 }
